+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.beta4 2015-??-??
+ - 2015-08-28 Fixed bug#[8146](http://bugs.otrs.org/show_bug.cgi?id=8146) - Generate report in Chinese from command line.
 
 #5.0.0.beta3 2015-09-01
  - 2015-08-25 Followup for bug#[9460](http://bugs.otrs.org/show_bug.cgi?id=9460) - Under some circumstances OTRS does not join Tickets to the Customernumber.